+++ mutations/css_chevron_inline.pl
@@ -1,0 +1,1 @@
+s/(\.compass-group-chevron \{\n)    display: inline-flex;\n    align-items: center;\n/$1/;